requirements.txt
mysql-connector-python
matplotlib

